+++ packages/compiler-cli/src/ngtsc/annotations/common/src/api.ts
