+++ packages/compiler-cli/src/ngtsc/annotations/common/src/factory.ts
@@ -8,7 +8,7 @@
 
 import {compileDeclareFactoryFunction, compileFactoryFunction, R3FactoryMetadata} from '@angular/compiler';
 
-import {CompileResult} from '../../transform';
+import {CompileResult} from '../../../transform';
 
 export type CompileFactoryFn = (metadata: R3FactoryMetadata) => CompileResult;
