+++ packages/compiler-cli/src/ngtsc/annotations/common/src/references_registry.ts
@@ -6,8 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Reference} from '../../imports';
-import {DeclarationNode} from '../../reflection';
+import {Reference} from '../../../imports';
+import {DeclarationNode} from '../../../reflection';
 
 /**
  * Implement this interface if you want DecoratorHandlers to register
