+++ packages/compiler-cli/src/ngtsc/annotations/common/src/util.ts
@@ -6,105 +6,14 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Expression, ExternalExpr, FactoryTarget, LiteralExpr, ParseLocation, ParseSourceFile, ParseSourceSpan, R3CompiledExpression, R3DependencyMetadata, R3FactoryMetadata, R3Reference, ReadPropExpr, Statement, WrappedNodeExpr} from '@angular/compiler';
+import {Expression, ExternalExpr, FactoryTarget, ParseLocation, ParseSourceFile, ParseSourceSpan, R3CompiledExpression, R3FactoryMetadata, R3Reference, ReadPropExpr, Statement, WrappedNodeExpr} from '@angular/compiler';
 import ts from 'typescript';
 
-import {ErrorCode, FatalDiagnosticError, makeDiagnostic, makeRelatedInformation} from '../../diagnostics';
-import {assertSuccessfulReferenceEmit, ImportFlags, Reference, ReferenceEmitter} from '../../imports';
-import {attachDefaultImportDeclaration} from '../../imports/src/default';
-import {ForeignFunctionResolver, PartialEvaluator} from '../../partial_evaluator';
-import {ClassDeclaration, CtorParameter, Decorator, Import, ImportedTypeValueReference, isNamedClassDeclaration, LocalTypeValueReference, ReflectionHost, TypeValueReference, TypeValueReferenceKind, UnavailableValue, ValueUnavailableKind} from '../../reflection';
-import {DeclarationData} from '../../scope';
-import {CompileResult} from '../../transform';
-
-export type ConstructorDeps = {
-  deps: R3DependencyMetadata[];
-}|{
-  deps: null;
-  errors: ConstructorDepError[];
-};
-
-export interface ConstructorDepError {
-  index: number;
-  param: CtorParameter;
-  reason: UnavailableValue;
-}
-
-export function getConstructorDependencies(
-    clazz: ClassDeclaration, reflector: ReflectionHost, isCore: boolean): ConstructorDeps|null {
-  const deps: R3DependencyMetadata[] = [];
-  const errors: ConstructorDepError[] = [];
-  let ctorParams = reflector.getConstructorParameters(clazz);
-  if (ctorParams === null) {
-    if (reflector.hasBaseClass(clazz)) {
-      return null;
-    } else {
-      ctorParams = [];
-    }
-  }
-  ctorParams.forEach((param, idx) => {
-    let token = valueReferenceToExpression(param.typeValueReference);
-    let attributeNameType: Expression|null = null;
-    let optional = false, self = false, skipSelf = false, host = false;
-
-    (param.decorators || []).filter(dec => isCore || isAngularCore(dec)).forEach(dec => {
-      const name = isCore || dec.import === null ? dec.name : dec.import!.name;
-      if (name === 'Inject') {
-        if (dec.args === null || dec.args.length !== 1) {
-          throw new FatalDiagnosticError(
-              ErrorCode.DECORATOR_ARITY_WRONG, Decorator.nodeForError(dec),
-              `Unexpected number of arguments to @Inject().`);
-        }
-        token = new WrappedNodeExpr(dec.args[0]);
-      } else if (name === 'Optional') {
-        optional = true;
-      } else if (name === 'SkipSelf') {
-        skipSelf = true;
-      } else if (name === 'Self') {
-        self = true;
-      } else if (name === 'Host') {
-        host = true;
-      } else if (name === 'Attribute') {
-        if (dec.args === null || dec.args.length !== 1) {
-          throw new FatalDiagnosticError(
-              ErrorCode.DECORATOR_ARITY_WRONG, Decorator.nodeForError(dec),
-              `Unexpected number of arguments to @Attribute().`);
-        }
-        const attributeName = dec.args[0];
-        token = new WrappedNodeExpr(attributeName);
-        if (ts.isStringLiteralLike(attributeName)) {
-          attributeNameType = new LiteralExpr(attributeName.text);
-        } else {
-          attributeNameType =
-              new WrappedNodeExpr(ts.createKeywordTypeNode(ts.SyntaxKind.UnknownKeyword));
-        }
-      } else {
-        throw new FatalDiagnosticError(
-            ErrorCode.DECORATOR_UNEXPECTED, Decorator.nodeForError(dec),
-            `Unexpected decorator ${name} on parameter.`);
-      }
-    });
-
-    if (token === null) {
-      if (param.typeValueReference.kind !== TypeValueReferenceKind.UNAVAILABLE) {
-        throw new Error(
-            'Illegal state: expected value reference to be unavailable if no token is present');
-      }
-      errors.push({
-        index: idx,
-        param,
-        reason: param.typeValueReference.reason,
-      });
-    } else {
-      deps.push({token, attributeNameType, optional, self, skipSelf, host});
-    }
-  });
-  if (errors.length === 0) {
-    return {deps};
-  } else {
-    return {deps: null, errors};
-  }
-}
+import {assertSuccessfulReferenceEmit, ImportedFile, ImportFlags, ModuleResolver, Reference, ReferenceEmitter} from '../../../imports';
+import {attachDefaultImportDeclaration} from '../../../imports/src/default';
+import {ForeignFunctionResolver, PartialEvaluator} from '../../../partial_evaluator';
+import {ClassDeclaration, Decorator, Import, ImportedTypeValueReference, isNamedClassDeclaration, LocalTypeValueReference, ReflectionHost, TypeValueReference, TypeValueReferenceKind} from '../../../reflection';
+import {CompileResult} from '../../../transform';
 
 /**
  * Convert a `TypeValueReference` to an `Expression` which refers to the type as a value.
@@ -135,125 +44,6 @@ export function valueReferenceToExpression(valueRef: TypeValueReference): Expres
     }
     return importExpr;
   }
-}
-
-/**
- * Convert `ConstructorDeps` into the `R3DependencyMetadata` array for those deps if they're valid,
- * or into an `'invalid'` signal if they're not.
- *
- * This is a companion function to `validateConstructorDependencies` which accepts invalid deps.
- */
-export function unwrapConstructorDependencies(deps: ConstructorDeps|null): R3DependencyMetadata[]|
-    'invalid'|null {
-  if (deps === null) {
-    return null;
-  } else if (deps.deps !== null) {
-    // These constructor dependencies are valid.
-    return deps.deps;
-  } else {
-    // These deps are invalid.
-    return 'invalid';
-  }
-}
-
-export function getValidConstructorDependencies(
-    clazz: ClassDeclaration, reflector: ReflectionHost, isCore: boolean): R3DependencyMetadata[]|
-    null {
-  return validateConstructorDependencies(
-      clazz, getConstructorDependencies(clazz, reflector, isCore));
-}
-
-/**
- * Validate that `ConstructorDeps` does not have any invalid dependencies and convert them into the
- * `R3DependencyMetadata` array if so, or raise a diagnostic if some deps are invalid.
- *
- * This is a companion function to `unwrapConstructorDependencies` which does not accept invalid
- * deps.
- */
-export function validateConstructorDependencies(
-    clazz: ClassDeclaration, deps: ConstructorDeps|null): R3DependencyMetadata[]|null {
-  if (deps === null) {
-    return null;
-  } else if (deps.deps !== null) {
-    return deps.deps;
-  } else {
-    // TODO(alxhub): this cast is necessary because the g3 typescript version doesn't narrow here.
-    // There is at least one error.
-    const error = (deps as {errors: ConstructorDepError[]}).errors[0];
-    throw createUnsuitableInjectionTokenError(clazz, error);
-  }
-}
-
-/**
- * Creates a fatal error with diagnostic for an invalid injection token.
- * @param clazz The class for which the injection token was unavailable.
- * @param error The reason why no valid injection token is available.
- */
-function createUnsuitableInjectionTokenError(
-    clazz: ClassDeclaration, error: ConstructorDepError): FatalDiagnosticError {
-  const {param, index, reason} = error;
-  let chainMessage: string|undefined = undefined;
-  let hints: ts.DiagnosticRelatedInformation[]|undefined = undefined;
-  switch (reason.kind) {
-    case ValueUnavailableKind.UNSUPPORTED:
-      chainMessage = 'Consider using the @Inject decorator to specify an injection token.';
-      hints = [
-        makeRelatedInformation(reason.typeNode, 'This type is not supported as injection token.'),
-      ];
-      break;
-    case ValueUnavailableKind.NO_VALUE_DECLARATION:
-      chainMessage = 'Consider using the @Inject decorator to specify an injection token.';
-      hints = [
-        makeRelatedInformation(
-            reason.typeNode,
-            'This type does not have a value, so it cannot be used as injection token.'),
-      ];
-      if (reason.decl !== null) {
-        hints.push(makeRelatedInformation(reason.decl, 'The type is declared here.'));
-      }
-      break;
-    case ValueUnavailableKind.TYPE_ONLY_IMPORT:
-      chainMessage =
-          'Consider changing the type-only import to a regular import, or use the @Inject decorator to specify an injection token.';
-      hints = [
-        makeRelatedInformation(
-            reason.typeNode,
-            'This type is imported using a type-only import, which prevents it from being usable as an injection token.'),
-        makeRelatedInformation(reason.node, 'The type-only import occurs here.'),
-      ];
-      break;
-    case ValueUnavailableKind.NAMESPACE:
-      chainMessage = 'Consider using the @Inject decorator to specify an injection token.';
-      hints = [
-        makeRelatedInformation(
-            reason.typeNode,
-            'This type corresponds with a namespace, which cannot be used as injection token.'),
-        makeRelatedInformation(reason.importClause, 'The namespace import occurs here.'),
-      ];
-      break;
-    case ValueUnavailableKind.UNKNOWN_REFERENCE:
-      chainMessage = 'The type should reference a known declaration.';
-      hints = [makeRelatedInformation(reason.typeNode, 'This type could not be resolved.')];
-      break;
-    case ValueUnavailableKind.MISSING_TYPE:
-      chainMessage =
-          'Consider adding a type to the parameter or use the @Inject decorator to specify an injection token.';
-      break;
-  }
-
-  const chain: ts.DiagnosticMessageChain = {
-    messageText: `No suitable injection token for parameter '${param.name || index}' of class '${
-        clazz.name.text}'.`,
-    category: ts.DiagnosticCategory.Error,
-    code: 0,
-    next: [{
-      messageText: chainMessage,
-      category: ts.DiagnosticCategory.Message,
-      code: 0,
-    }],
-  };
-
-  return new FatalDiagnosticError(ErrorCode.PARAM_MISSING_TOKEN, param.nameNode, chain, hints);
 }
 
 export function toR3Reference(
@@ -458,35 +248,6 @@ export function wrapFunctionExpressionsInParens(expression: ts.Expression): ts.E
 }
 
 /**
- * Create a `ts.Diagnostic` which indicates the given class is part of the declarations of two or
- * more NgModules.
- *
- * The resulting `ts.Diagnostic` will have a context entry for each NgModule showing the point where
- * the directive/pipe exists in its `declarations` (if possible).
- */
-export function makeDuplicateDeclarationError(
-    node: ClassDeclaration, data: DeclarationData[], kind: string): ts.Diagnostic {
-  const context: ts.DiagnosticRelatedInformation[] = [];
-  for (const decl of data) {
-    if (decl.rawDeclarations === null) {
-      continue;
-    }
-    // Try to find the reference to the declaration within the declarations array, to hang the
-    // error there. If it can't be found, fall back on using the NgModule's name.
-    const contextNode = decl.ref.getOriginForDiagnostics(decl.rawDeclarations, decl.ngModule.name);
-    context.push(makeRelatedInformation(
-        contextNode,
-        `'${node.name.text}' is listed in the declarations of the NgModule '${
-            decl.ngModule.name.text}'.`));
-  }
-
-  // Finally, produce the diagnostic.
-  return makeDiagnostic(
-      ErrorCode.NGMODULE_DECLARATION_NOT_UNIQUE, node.name,
-      `The ${kind} '${node.name.text}' is declared by more than one NgModule.`, context);
-}
-
-/**
  * Resolves the given `rawProviders` into `ClassDeclarations` and returns
  * a set containing those that are known to require a factory definition.
  * @param rawProviders Expression that declared the providers array in the source.
@@ -595,4 +356,24 @@ export function toFactoryMetadata(
     deps: meta.deps,
     target
   };
+}
+
+export function resolveImportedFile(
+    moduleResolver: ModuleResolver, importedFile: ImportedFile, expr: Expression,
+    origin: ts.SourceFile): ts.SourceFile|null {
+  // If `importedFile` is not 'unknown' then it accurately reflects the source file that is
+  // being imported.
+  if (importedFile !== 'unknown') {
+    return importedFile;
+  }
+
+  // Otherwise `expr` has to be inspected to determine the file that is being imported. If `expr`
+  // is not an `ExternalExpr` then it does not correspond with an import, so return null in that
+  // case.
+  if (!(expr instanceof ExternalExpr)) {
+    return null;
+  }
+
+  // Figure out what file is being imported.
+  return moduleResolver.resolveModule(expr.value.moduleName!, origin.fileName);
 }
