+++ packages/compiler-cli/src/ngtsc/annotations/component/test/component_spec.ts
@@ -9,19 +9,19 @@
 import {ConstantPool} from '@angular/compiler';
 import ts from 'typescript';
 
-import {CycleAnalyzer, CycleHandlingStrategy, ImportGraph} from '../../cycles';
-import {ErrorCode, FatalDiagnosticError} from '../../diagnostics';
-import {absoluteFrom} from '../../file_system';
-import {runInEachFileSystem} from '../../file_system/testing';
-import {ModuleResolver, Reference, ReferenceEmitter} from '../../imports';
-import {CompoundMetadataReader, DtsMetadataReader, InjectableClassRegistry, LocalMetadataRegistry, ResourceRegistry} from '../../metadata';
-import {PartialEvaluator} from '../../partial_evaluator';
-import {NOOP_PERF_RECORDER} from '../../perf';
-import {isNamedClassDeclaration, TypeScriptReflectionHost} from '../../reflection';
-import {LocalModuleScopeRegistry, MetadataDtsModuleScopeResolver, TypeCheckScopeRegistry} from '../../scope';
-import {getDeclaration, makeProgram} from '../../testing';
-import {ResourceLoader, ResourceLoaderContext} from '../src/api';
-import {ComponentDecoratorHandler} from '../src/component';
+import {CycleAnalyzer, CycleHandlingStrategy, ImportGraph} from '../../../cycles';
+import {ErrorCode, FatalDiagnosticError} from '../../../diagnostics';
+import {absoluteFrom} from '../../../file_system';
+import {runInEachFileSystem} from '../../../file_system/testing';
+import {ModuleResolver, Reference, ReferenceEmitter} from '../../../imports';
+import {CompoundMetadataReader, DtsMetadataReader, InjectableClassRegistry, LocalMetadataRegistry, ResourceRegistry} from '../../../metadata';
+import {PartialEvaluator} from '../../../partial_evaluator';
+import {NOOP_PERF_RECORDER} from '../../../perf';
+import {isNamedClassDeclaration, TypeScriptReflectionHost} from '../../../reflection';
+import {LocalModuleScopeRegistry, MetadataDtsModuleScopeResolver, TypeCheckScopeRegistry} from '../../../scope';
+import {getDeclaration, makeProgram} from '../../../testing';
+import {ResourceLoader, ResourceLoaderContext} from '../../common';
+import {ComponentDecoratorHandler} from '../src/handler';
 
 export class StubResourceLoader implements ResourceLoader {
   resolve(v: string): string {
@@ -76,7 +76,7 @@ function setup(program: ts.Program, options: ts.CompilerOptions, host: ts.Compil
       /* i18nUseExternalIds */ true,
       /* enableI18nLegacyMessageIdFormat */ false,
       /* usePoisonedData */ false,
-      /* i18nNormalizeLineEndingsInICUs */ undefined,
+      /* i18nNormalizeLineEndingsInICUs */ false,
       moduleResolver,
       cycleAnalyzer,
       CycleHandlingStrategy.UseRemoteScoping,
