+++ packages/compiler-cli/src/ngtsc/annotations/directive/BUILD.bazel
@@ -3,35 +3,22 @@ load("//tools:defaults.bzl", "ts_library")
 package(default_visibility = ["//visibility:public"])
 
 ts_library(
-    name = "annotations",
+    name = "directive",
     srcs = ["index.ts"] + glob([
         "src/**/*.ts",
     ]),
     deps = [
         "//packages/compiler",
         "//packages/compiler-cli/src/ngtsc/annotations/common",
-        "//packages/compiler-cli/src/ngtsc/annotations/component",
-        "//packages/compiler-cli/src/ngtsc/annotations/directive",
-        "//packages/compiler-cli/src/ngtsc/annotations/ng_module",
-        "//packages/compiler-cli/src/ngtsc/cycles",
         "//packages/compiler-cli/src/ngtsc/diagnostics",
-        "//packages/compiler-cli/src/ngtsc/file_system",
         "//packages/compiler-cli/src/ngtsc/imports",
-        "//packages/compiler-cli/src/ngtsc/incremental:api",
         "//packages/compiler-cli/src/ngtsc/incremental/semantic_graph",
-        "//packages/compiler-cli/src/ngtsc/indexer",
         "//packages/compiler-cli/src/ngtsc/metadata",
         "//packages/compiler-cli/src/ngtsc/partial_evaluator",
         "//packages/compiler-cli/src/ngtsc/perf",
         "//packages/compiler-cli/src/ngtsc/reflection",
         "//packages/compiler-cli/src/ngtsc/scope",
-        "//packages/compiler-cli/src/ngtsc/shims:api",
         "//packages/compiler-cli/src/ngtsc/transform",
-        "//packages/compiler-cli/src/ngtsc/typecheck/api",
-        "//packages/compiler-cli/src/ngtsc/typecheck/diagnostics",
-        "//packages/compiler-cli/src/ngtsc/typecheck/extended/api",
-        "//packages/compiler-cli/src/ngtsc/util",
-        "//packages/compiler-cli/src/ngtsc/xi18n",
         "@npm//@types/node",
         "@npm//typescript",
     ],
