+++ packages/compiler-cli/src/ngtsc/annotations/directive/README.md
@@ -1,0 +1,5 @@
+# What is the 'annotations/directive' package?
+
+This package implements the `DirectiveDecoratorHandler`, which processes and compiles `@Directive`-decorated classes.
+
+Directive compilation is also the base of component compilation, and so this package exports a number of utilities related to directive compilation for the 'annotations/component' package to use.
