+++ packages/compiler-cli/src/ngtsc/annotations/directive/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+export {DirectiveDecoratorHandler} from './src/handler';
+export {DirectiveSymbol} from './src/symbol';
+export * from './src/shared';
