+++ packages/compiler-cli/src/ngtsc/annotations/directive/test/directive_spec.ts
@@ -8,16 +8,16 @@
 import {CssSelector, DirectiveMeta as T2DirectiveMeta, parseTemplate, R3TargetBinder, SelectorMatcher, TmplAstElement} from '@angular/compiler';
 import ts from 'typescript';
 
-import {absoluteFrom} from '../../file_system';
-import {runInEachFileSystem} from '../../file_system/testing';
-import {ReferenceEmitter} from '../../imports';
-import {DtsMetadataReader, InjectableClassRegistry, LocalMetadataRegistry} from '../../metadata';
-import {PartialEvaluator} from '../../partial_evaluator';
-import {NOOP_PERF_RECORDER} from '../../perf';
-import {ClassDeclaration, isNamedClassDeclaration, TypeScriptReflectionHost} from '../../reflection';
-import {LocalModuleScopeRegistry, MetadataDtsModuleScopeResolver} from '../../scope';
-import {getDeclaration, makeProgram} from '../../testing';
-import {DirectiveDecoratorHandler} from '../src/directive';
+import {absoluteFrom} from '../../../file_system';
+import {runInEachFileSystem} from '../../../file_system/testing';
+import {ReferenceEmitter} from '../../../imports';
+import {DtsMetadataReader, InjectableClassRegistry, LocalMetadataRegistry} from '../../../metadata';
+import {PartialEvaluator} from '../../../partial_evaluator';
+import {NOOP_PERF_RECORDER} from '../../../perf';
+import {ClassDeclaration, isNamedClassDeclaration, TypeScriptReflectionHost} from '../../../reflection';
+import {LocalModuleScopeRegistry, MetadataDtsModuleScopeResolver} from '../../../scope';
+import {getDeclaration, makeProgram} from '../../../testing';
+import {DirectiveDecoratorHandler} from '../index';
 
 runInEachFileSystem(() => {
   let _: typeof absoluteFrom;
