+++ packages/compiler-cli/src/ngtsc/annotations/ng_module/BUILD.bazel
@@ -3,43 +3,26 @@ load("//tools:defaults.bzl", "ts_library")
 package(default_visibility = ["//visibility:public"])
 
 ts_library(
-    name = "ngcc",
-    srcs = glob([
-        "*.ts",
-        "**/*.ts",
+    name = "ng_module",
+    srcs = ["index.ts"] + glob([
+        "src/**/*.ts",
     ]),
-    tsconfig = "//packages/compiler-cli:tsconfig",
     deps = [
-        "//packages:types",
         "//packages/compiler",
-        "//packages/compiler-cli",
-        "//packages/compiler-cli:import_meta_url_types",
-        "//packages/compiler-cli/src/ngtsc/annotations",
         "//packages/compiler-cli/src/ngtsc/annotations/common",
-        "//packages/compiler-cli/src/ngtsc/cycles",
         "//packages/compiler-cli/src/ngtsc/diagnostics",
-        "//packages/compiler-cli/src/ngtsc/file_system",
         "//packages/compiler-cli/src/ngtsc/imports",
         "//packages/compiler-cli/src/ngtsc/incremental:api",
         "//packages/compiler-cli/src/ngtsc/incremental/semantic_graph",
-        "//packages/compiler-cli/src/ngtsc/logging",
         "//packages/compiler-cli/src/ngtsc/metadata",
         "//packages/compiler-cli/src/ngtsc/partial_evaluator",
         "//packages/compiler-cli/src/ngtsc/perf",
         "//packages/compiler-cli/src/ngtsc/reflection",
         "//packages/compiler-cli/src/ngtsc/scope",
-        "//packages/compiler-cli/src/ngtsc/sourcemaps",
+        "//packages/compiler-cli/src/ngtsc/shims:api",
         "//packages/compiler-cli/src/ngtsc/transform",
-        "//packages/compiler-cli/src/ngtsc/translator",
         "//packages/compiler-cli/src/ngtsc/util",
-        "@npm//@types/convert-source-map",
         "@npm//@types/node",
-        "@npm//@types/semver",
-        "@npm//@types/yargs",
-        "@npm//dependency-graph",
-        "@npm//magic-string",
-        "@npm//semver",
         "@npm//typescript",
-        "@npm//yargs",
     ],
 )
