+++ packages/compiler-cli/src/ngtsc/annotations/ng_module/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+export {NgModuleDecoratorHandler, NgModuleSymbol} from './src/handler';
