+++ packages/compiler-cli/src/ngtsc/annotations/ng_module/test/ng_module_spec.ts
@@ -9,17 +9,17 @@ import {WrappedNodeExpr} from '@angular/compiler';
 import {R3Reference} from '@angular/compiler/src/compiler';
 import ts from 'typescript';
 
-import {absoluteFrom} from '../../file_system';
-import {runInEachFileSystem} from '../../file_system/testing';
-import {LocalIdentifierStrategy, ReferenceEmitter} from '../../imports';
-import {CompoundMetadataReader, DtsMetadataReader, InjectableClassRegistry, LocalMetadataRegistry} from '../../metadata';
-import {PartialEvaluator} from '../../partial_evaluator';
-import {NOOP_PERF_RECORDER} from '../../perf';
-import {isNamedClassDeclaration, TypeScriptReflectionHost} from '../../reflection';
-import {LocalModuleScopeRegistry, MetadataDtsModuleScopeResolver} from '../../scope';
-import {getDeclaration, makeProgram} from '../../testing';
-import {NgModuleDecoratorHandler} from '../src/ng_module';
-import {NoopReferencesRegistry} from '../src/references_registry';
+import {absoluteFrom} from '../../../file_system';
+import {runInEachFileSystem} from '../../../file_system/testing';
+import {LocalIdentifierStrategy, ReferenceEmitter} from '../../../imports';
+import {CompoundMetadataReader, DtsMetadataReader, InjectableClassRegistry, LocalMetadataRegistry} from '../../../metadata';
+import {PartialEvaluator} from '../../../partial_evaluator';
+import {NOOP_PERF_RECORDER} from '../../../perf';
+import {isNamedClassDeclaration, TypeScriptReflectionHost} from '../../../reflection';
+import {LocalModuleScopeRegistry, MetadataDtsModuleScopeResolver} from '../../../scope';
+import {getDeclaration, makeProgram} from '../../../testing';
+import {NoopReferencesRegistry} from '../../common';
+import {NgModuleDecoratorHandler} from '../src/handler';
 
 runInEachFileSystem(() => {
   describe('NgModuleDecoratorHandler', () => {
